refactor(layout): make index files and separate files for overlay layout
- move types/interfaces into overlays/types.ts
- add overlays/index.ts
- add overlays/layoutStyles
- add utils.ts - and move stringifyOverlayKeysType and parseOverlayKeysType there
- add index.ts to head layout folder
