+++ frontend/src/layout/index.ts
@@ -1,0 +1,7 @@
+export {
+  initialLayoutWidgets,
+  initialToolboxWidgets,
+  widgetsKeys,
+} from "./initialLayoutWidgets";
+
+export * from "./overlays";
